+++ release_notes.md
@@ -6,4 +6,6 @@
 - Add JitTrace Files for v4.1042
 - Updating OTel related nuget packages (#11216)
 - Moving to version 1.5.7 of Microsoft.Azure.AppService.Middleware.Functions (#11231)
+- Refactor code to move the logic to search for WorkerConfigs to a default worker configuration resolver (#11219)
+- Update Node.js Worker Version to [3.11.0](https://github.com/Azure/azure-functions-nodejs-worker/releases/tag/v3.11.0)
 - Update outdated bundle message - removed deprecation date (#11230)